refactor(backend): extract endpoint crud module
Continues the incremental split of the crud monolith. The single-item
Endpoint CRUD functions move into the existing crud/endpoint.py, which
already held bulk_delete_endpoints; callers switch to direct submodule
imports. Pure move -- no logic changes.
