+++ apps/backend/src/rhesis/backend/app/crud/__init__.py
@@ -26,83 +26,6 @@ from rhesis.backend.app.utils.query_utils import QueryBuilder, include
 logger = logging.getLogger(__name__)
 
 
-# Endpoint CRUD
-_ENDPOINT_RELATED_FIELDS = (
-    include(models.Endpoint.status),
-    include(models.Endpoint.user),
-    include(models.Endpoint.project),
-)
-
-
-def get_endpoint(
-    db: Session,
-    endpoint_id: uuid.UUID,
-    organization_id: str,
-    user_id: str,
-    project_id: str = None,
-) -> Optional[models.Endpoint]:
-    """Get endpoint with relationships eagerly loaded."""
-    return get_item_detail(
-        db,
-        models.Endpoint,
-        endpoint_id,
-        organization_id,
-        user_id,
-        project_id=project_id,
-        related_fields=_ENDPOINT_RELATED_FIELDS,
-    )
-
-
-def get_endpoints(
-    db: Session,
-    skip: int = 0,
-    limit: int = 10,
-    sort_by: str = "created_at",
-    sort_order: str = "desc",
-    filter: str | None = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> List[models.Endpoint]:
-    return get_items_detail(
-        db,
-        models.Endpoint,
-        skip,
-        limit,
-        sort_by,
-        sort_order,
-        filter,
-        related_fields=_ENDPOINT_RELATED_FIELDS,
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-
-
-def create_endpoint(
-    db: Session, endpoint: schemas.EndpointCreate, organization_id: str, user_id: str
-) -> models.Endpoint:
-    """Create endpoint."""
-    return create_item(db, models.Endpoint, endpoint, organization_id, user_id)
-
-
-def update_endpoint(
-    db: Session,
-    endpoint_id: uuid.UUID,
-    endpoint: schemas.EndpointUpdate,
-    organization_id: str,
-    user_id: str,
-) -> Optional[models.Endpoint]:
-    """Update endpoint."""
-    return update_item(db, models.Endpoint, endpoint_id, endpoint, organization_id, user_id)
-
-
-def delete_endpoint(
-    db: Session, endpoint_id: uuid.UUID, organization_id: str, user_id: str
-) -> Optional[models.Endpoint]:
-    return delete_item(
-        db, models.Endpoint, endpoint_id, organization_id=organization_id, user_id=user_id
-    )
-
-
 # TestSet CRUD
 def get_test_set(
     db: Session, test_set_id: uuid.UUID, organization_id: str = None, user_id: str = None
