+++ tests/backend/routes/test_endpoint_explore.py
@@ -47,10 +47,10 @@ class TestExploreEndpointHappy:
         mock_async_result.id = str(uuid.uuid4())
 
         with (
-            patch("rhesis.backend.app.routers.endpoint.crud") as mock_crud,
+            patch("rhesis.backend.app.routers.endpoint.endpoint_crud") as mock_endpoint_crud,
             patch("rhesis.backend.app.routers.endpoint.launch_job", return_value=mock_async_result),
         ):
-            mock_crud.get_endpoint.return_value = mock_ep
+            mock_endpoint_crud.get_endpoint.return_value = mock_ep
 
             response = app.post(
                 f"/endpoints/{endpoint_id}/explore",
@@ -68,10 +68,10 @@ class TestExploreEndpointHappy:
         mock_async_result.id = str(uuid.uuid4())
 
         with (
-            patch("rhesis.backend.app.routers.endpoint.crud") as mock_crud,
+            patch("rhesis.backend.app.routers.endpoint.endpoint_crud") as mock_endpoint_crud,
             patch("rhesis.backend.app.routers.endpoint.launch_job", return_value=mock_async_result),
         ):
-            mock_crud.get_endpoint.return_value = mock_ep
+            mock_endpoint_crud.get_endpoint.return_value = mock_ep
 
             response = app.post(
                 f"/endpoints/{endpoint_id}/explore",
@@ -88,8 +88,8 @@ class TestExploreEndpointHappy:
 
 class TestExploreEndpoint404:
     def test_returns_404_for_unknown_endpoint(self, app, endpoint_id):
-        with patch("rhesis.backend.app.routers.endpoint.crud") as mock_crud:
-            mock_crud.get_endpoint.return_value = None
+        with patch("rhesis.backend.app.routers.endpoint.endpoint_crud") as mock_endpoint_crud:
+            mock_endpoint_crud.get_endpoint.return_value = None
 
             response = app.post(
                 f"/endpoints/{endpoint_id}/explore",
