+++ tests/backend/security/test_organization_filtering.py
@@ -11,6 +11,7 @@ import pytest
 from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models
+from rhesis.backend.app.crud import endpoint as endpoint_crud
 from rhesis.backend.app.crud import model as model_crud
 from rhesis.backend.app.crud import prompt as prompt_crud
 from rhesis.backend.app.crud import task as task_crud
@@ -163,7 +164,7 @@ class TestCrudOrganizationFiltering:
         # Create an endpoint first (required for test configuration)
         endpoint_data = EndpointDataFactory.minimal_data()
         endpoint_data["project_id"] = str(project.id)
-        endpoint = crud.create_endpoint(
+        endpoint = endpoint_crud.create_endpoint(
             db=test_db, endpoint=endpoint_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
@@ -253,7 +254,7 @@ class TestCrudOrganizationFiltering:
         # Create an endpoint first (required for test configuration)
         endpoint_data = EndpointDataFactory.minimal_data()
         endpoint_data["project_id"] = str(project.id)
-        endpoint = crud.create_endpoint(
+        endpoint = endpoint_crud.create_endpoint(
             db=test_db, endpoint=endpoint_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
@@ -340,12 +341,12 @@ class TestCrudOrganizationFiltering:
             config_source="manual",
             project_id=project.id,
         )
-        endpoint = crud.create_endpoint(
+        endpoint = endpoint_crud.create_endpoint(
             db=test_db, endpoint=endpoint_data, organization_id=str(org1.id), user_id=str(user1.id)
         )
 
         # User from org1 should be able to access the endpoint
-        result_org1 = crud.get_endpoint(
+        result_org1 = endpoint_crud.get_endpoint(
             test_db, endpoint.id, organization_id=str(org1.id), user_id=str(user1.id)
         )
         assert result_org1 is not None
@@ -353,7 +354,7 @@ class TestCrudOrganizationFiltering:
         assert str(result_org1.organization_id) == str(org1.id)
 
         # User from org2 should NOT be able to access the endpoint
-        result_org2 = crud.get_endpoint(
+        result_org2 = endpoint_crud.get_endpoint(
             test_db, endpoint.id, organization_id=str(org2.id), user_id=str(user2.id)
         )
         assert result_org2 is None
@@ -613,7 +614,7 @@ class TestCrudParameterValidation:
             ("get_test", crud.get_test),
             ("get_test_result", test_result_crud.get_test_result),
             ("get_test_run", get_test_run),
-            ("get_endpoint", crud.get_endpoint),
+            ("get_endpoint", endpoint_crud.get_endpoint),
             ("get_prompt", prompt_crud.get_prompt),
             ("get_model", model_crud.get_model),
             ("get_metric", get_metric),
